requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
cryptography==46.0.2
cssselect2==0.8.0
Django==5.2.7
django-allauth==65.12.0
django-auditlog==3.3.0
django-crispy-forms==2.4
django-embed-video==1.4.10
django-environ==0.12.0
django-formtools==2.5.1
django-fsm==3.0.1
django-fsm-admin-v4==1.2.5
django-multiselectfield==1.0.1
django-tenants==3.9.0
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
Faker==37.11.0
fonttools==4.60.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
kombu==5.5.4
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
redis==5.2.1
reportlab==4.4.5
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==53.3
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.11.0
zopfli==0.2.3.post1
